// src/models/survey-elements/closedQuestion.ts
export interface ClosedQuestion {
    cId: number;
    elementId: number;
    elementTitle: string;
    situation: string;
    questiontext: string;
    answer1: string;
    answer2: string;
    answer3: string;
    answer4: string;
    answer5: string;
    answer6: string;
    optionalTextfield: boolean;
    multipleSelection: boolean;
    picture: string;
    surveyId: number;
}